+++ pkg/config/criteria.go
@@ -33,6 +33,19 @@ var allowedAttributeNames = map[string]struct{}{
 	AttrOwnerName:       {},
 }
 
+type Config struct {
+	// Discovery configuration
+	Discovery DiscoveryConfig `yaml:"discovery"`
+}
+
+func (c *Config) Validate() error {
+	if err := c.Discovery.Services.Validate(); err != nil {
+		return fmt.Errorf("error in services YAML property: %w", err)
+	}
+
+	return nil
+}
+
 // DiscoveryConfig is the configuration.
 type DiscoveryConfig struct {
 	// Services selection.
